oppgrader til eslint 9
