--- .eslintrc.cjs
@@ -1,27 +1,0 @@
-const IGNORED_UNICORN_RULES = {
-    "unicorn/filename-case": "off",
-};
-
-// eslint-disable-next-line unicorn/prefer-module,no-undef
-module.exports = {
-    parser: "@typescript-eslint/parser",
-    env: {
-        es2022: true,
-        node: true,
-    },
-    extends: ["eslint:recommended", "plugin:@typescript-eslint/recommended", "prettier", "plugin:unicorn/recommended"],
-    overrides: [],
-    parserOptions: {
-        ecmaVersion: "latest",
-        sourceType: "module",
-    },
-    plugins: ["react", "prettier", "@typescript-eslint", "simple-import-sort"],
-    rules: {
-        eqeqeq: ["error", "always"],
-        "prettier/prettier": ["error"],
-        "import/prefer-default-export": "off",
-        "simple-import-sort/imports": "error",
-        "simple-import-sort/exports": "error",
-        ...IGNORED_UNICORN_RULES,
-    },
-};
