+++ eslint.config.mjs
@@ -1,0 +1,37 @@
+import eslint from "@eslint/js";
+import eslintPluginPrettierRecommended from "eslint-plugin-prettier/recommended";
+import simpleImportSort from "eslint-plugin-simple-import-sort";
+import eslintPluginUnicorn from "eslint-plugin-unicorn";
+import tseslint from "typescript-eslint";
+
+const IGNORED_UNICORN_RULES = {
+    "unicorn/filename-case": "off",
+    "unicorn/no-null": "off",
+    "unicorn/prevent-abbreviations": "off",
+    "unicorn/no-nested-ternary": "off",
+};
+
+export default tseslint.config(
+    {
+        ignores: ["dist/*"],
+    },
+    eslint.configs.recommended,
+    ...tseslint.configs.recommended,
+    {
+        plugins: {
+            "simple-import-sort": simpleImportSort,
+        },
+        rules: {
+            "simple-import-sort/imports": "error",
+            "simple-import-sort/exports": "error",
+        },
+    },
+    eslintPluginUnicorn.configs["flat/recommended"],
+    {
+        rules: {
+            eqeqeq: ["error", "always"],
+            ...IGNORED_UNICORN_RULES,
+        },
+    },
+    eslintPluginPrettierRecommended,
+);
