+++ package.json
@@ -16,9 +16,9 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "build": "tsc",
-    "check-types": "tsc --noemit",
-    "lint": "eslint --ext .js,.ts, .",
-    "lint:fix": "eslint --fix --ext .js,.ts, ."
+    "types:check": "tsc --noemit",
+    "lint": "eslint",
+    "lint:fix": "eslint --fix"
   },
   "peerDependencies": {
     "express": "4.x"
@@ -27,17 +27,17 @@
     "cookie-parser": "1.4.6"
   },
   "devDependencies": {
+    "@eslint/eslintrc": "3.1.0",
+    "@eslint/js": "9.9.1",
+    "@types/cookie-parser": "1.4.7",
     "@types/express": "4.17.21",
     "@types/node": "22.5.1",
-    "@typescript-eslint/eslint-plugin": "6.13.2",
-    "eslint": "8.55.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-import": "2.29.1",
+    "eslint": "9.9.1",
     "eslint-plugin-prettier": "5.2.1",
-    "eslint-plugin-react": "7.35.0",
-    "eslint-plugin-simple-import-sort": "10.0.0",
-    "eslint-plugin-unicorn": "49.0.0",
+    "eslint-plugin-simple-import-sort": "12.1.1",
+    "eslint-plugin-unicorn": "55.0.0",
     "prettier": "3.3.3",
-    "typescript": "5.5.4"
+    "typescript": "5.5.4",
+    "typescript-eslint": "8.3.0"
   }
 }
